refactor: components directory struct
