+++ src/pages/index.vue
@@ -10,7 +10,7 @@
 import { Component, mixins } from 'nuxt-property-decorator';
 import { StoreHelper, StoreHelperMixin, Actions } from '../mixins/store_helper';
 import * as ns from '../namespace_maps';
-import List from '../store_modules/todo/components/list.vue';
+import List from '../components/todo/list.vue';
 import { TodoActions } from '../store_modules/todo/actions';
 import { initialStateResolver } from '../store_modules/todo/index';
 import { Store } from 'vuex';
